Added Theme Service and navbar buttons to change theme
