+++ projects/swimlane/ngx-ui/src/lib/services/index.ts
@@ -1,3 +1,4 @@
 export * from './injection.service';
 export * from './injection-registery.service';
 export * from './icon-registery.service';
+export * from './theme.service';
