+++ projects/swimlane/ngx-ui/src/lib/utils/index.ts
@@ -3,3 +3,4 @@ export * from './throttle';
 export * from './id';
 export * from './position';
 export * from './keys';
+export * from './theme.enum';
