+++ projects/swimlane/ngx-ui/src/lib/utils/theme.enum.ts
@@ -1,0 +1,5 @@
+export enum ThemeEnum {
+  Day = 'day',
+  Night = 'night',
+  Moonlight = 'night'
+}
