+++ src/app/components/hotkeys-page/hotkeys-page.component.html
